requirements.txt
flask
gunicorn
newsapi-python
textblob
oauthlib
requests
flask_login
numpy
pandas
pycorenlp
psycopg2
json
time
datetime
os
hashlib